+++ example/example.dart
@@ -650,8 +650,7 @@ class JsonParser {
           state.position = $pos2;
         }
       }
-      final $val = state.substring(start, state.position);
-      final s = $val;
+      final s = state.substring(start, state.position);
       parseS(state);
       return Ok(flag && s.length <= 18 ? int.parse(s) : num.parse(s));
     } else {