+++ lib/src/parsers/text/text_parser.dart
@@ -350,8 +350,7 @@ class TextParser {
     }
     if ($list.isNotEmpty) {
       final n = $list;
-      final $val = SequenceExpression(expressions: n);
-      final e = $val;
+      final e = SequenceExpression(expressions: n);
       e.sourceCode = state.substring(pos, state.position).trimRight();
       return Ok(e);
     } else {
@@ -841,8 +840,7 @@ class TextParser {
         if (state.ch == 41) {
           state.nextChar();
           parseS(state);
-          final $val = GroupExpression(expression: e);
-          final g = $val;
+          final g = GroupExpression(expression: e);
           t != null ? g.type = t : null;
           return Ok(g);
         } else {